+++ src/solvers/hypre/Hypre.hpp
@@ -96,7 +96,7 @@ struct Hypre : public SparseSystem
   /// @param maxit       The maximum number of iterations.
   /// @param err         The error tolerance used during solving.
   Hypre(const PGFem3D_opt& options, MPI_Comm comm, const int Ap[],
-        const Ai_t Ai[], const long rowsPerProc[], long maxit, double err);
+        const Ai_t Ai[], const long rowsPerProc[], int maxit, double err);
 
   ~Hypre();
